Handle empty spaces between walls as OUTSIDE.
This prevents certain levels from showing FLOOR outside the board.
A flood fill algorithm starting at the position of the man is used, if the
board should contain FLOOR(not BOX_DESTINATION) in a non conjugate area, this
will not work, a solution would be to flood fill from every box, but that is
not nescessary with the current levels.